Add files via upload
This file has been renamed but was called MOTE2 in the bayesian code. Please use these data as a reference for how to create an excel file that is similar, or how to use quantification in histology. These are the signs I use, as you'll notice not all of them are in the bayesian models that I typically run, but the benefit of the data being in this style is how versatile they are. Some of these data are able to be correlated to gene expression data (see Beavers et al 2023) but are not necessarily used from my own analysis.